refactor(generate): extract the scoped-generation engine into core
Move the scoped page-generation engine out of the CLI so the OSS server and
other core consumers can drive it without importing the CLI package.

- core/generation/scope.py: the pure scope resolver (intent or ranked coverage
  into a plan of page ids to generate, dependents to mark stale, and a cost
  plan) moved from the CLI; the CLI module is now a re-export shim so existing
  imports and tests are unchanged.
- core/pipeline/reparse.py: reparse_repo(), the parse-without-graph step, with
  the submodule and nested-repo flags injected rather than read from state, so
  any caller can supply its own.
- core/pipeline/scoped_generation.py: rehydrate_repo() (rehydrate graph and git
  from SQL, re-parse off the event loop, load records, build the cascade
  dependency map) and execute_scoped_generation() (generate the requested
  subset, persist, mark dependents stale, heal related pages, index FTS). Both
  are injectable and free of CLI and server dependencies.

The CLI generate engine now composes these two calls, keeping its cost gate,
interactive chooser, and reporting. Behaviour is unchanged; the graph-metric
work is offloaded to a thread so an async caller never stalls its loop, and the
full-text index step is best-effort again.
